refactor: alert창 관련 로직 함수로 분리
